ChainstateManager class.
This introduces the ChainstateManager class, which provides an interface
to the global chain state (e.g. UTXO database or mapBlockIndex).

For now, the class has no state on its own and just provides access to
the underlying globals.  Converting all code to use the ChainstateManager
instead of the globals directly will be the first step in removing the
globals altogether.

Once that is done, we will be able to move the actual data into the
ChainstateManager instances, and pass them down through function calls
to all the places that need them.