Mark some CBlockIndex* usages as const.
Most CBlockIndex instances in the code are actually read-only (at least
after being validated and attached to a chain), while lots of code uses
non-const CBlockIndex* references to them.

This commit updates a lot of the places to use proper const-ness where
this is easily possible and trivial to do.  Some places also refactor
for-loops into foreach and use "auto" for iterators to generally
improve code readability but also make the code flexible with respect
to const-ness changes of e.g. vector/map value types.