Avoid pindex mutation during VerifyDB.
CVerifyDB is (at level 4) trying to reconnect the last N blocks of the
active chain; however, ConnectBlock may actually modify a CBlockIndex,
namely when updating its validity, undo data and calculated money supply.

We want to avoid that DB verification modifies the actual block index
instances in the active chain.  Thus this change modifies the code to
apply ConnectBlock on a temporary CBlockIndex copy, verifying that the
resulting values of the potentially-modified fields match what we expect.