Always use script signature cache.
The CachingTransactionSignatureChecker had a flag that allowed instances
to be configured to actually not use the cache, and that flag was passed
through all the way from e.g. ConnectBlock.

In practice, however, it was always set to true except for unit tests
and assert's (which might be completely compiled out) during mempool
validation.  (So in particular, it was and is true for all normal
control flow during operation.)

There is no valid reason for why signature caching should ever be turned
off (in particular not if it is just done in special cases and not during
normal operations).  This change removes the flag and simplifies the
corresponding code.