Introduce fJustCheck flag for DisconnectBlock.
Unlike ConnectBlock, DisconnectBlock does not have a direct flag for
disabling on-disk index updates (fJustCheck).  It has pfClean, which
acts in a similar way (when the flag isn't passed as null, then
index updates are disabled).  But this behaviour is hard to understand
and perhaps unexpected for callers.

This commit refactors the way the method works, by changing pfClean into
a clear and direct fJustCheck flag.  The previous "result" returned in
pfClean is now returned from DisconnectBlock directly.  The new flag is
used from VerifyDB similarly to how fJustCheck for ConnectBlock is used
from this place.