Singleton ChainstateManager instance.
Instead of having many ChainstateManager instances (which all refer to
the same global state in the end), this turns the class into a singleton.
This is enforced by assert's, and all places except for the one instance
during the init life cycle now use references to the singleton instead
of constructing their own instance.