Explicit ref counting for the ChainstateManager.
This introduces a ChainstateManager::Reference class, which replaces
the use of direct references (via ChainstateManager::Get) to the "global"
singleton instance.

These classes keep a refcount of the global instance, and we enforce that
all references are destructed before the global instance ends its life
from init.  This helps prevent potential race conditions and other
use-after-free issues, e.g. with multiple threads that are not properly
synchronised to init.

This measure will be temporary.  The final goal is to get rid of the
singleton-global anyway, and instead pass the one instance from init
explicity to all places that need it.