Proper life cycle for CoinMintingModule.
This refactors how the CoinMintingModule in miner.cpp is managed.
In particular, we now also destruct it when the node is stopped (not
just initialise it on startup), so that it has a proper life cycle
that can be synchronised with other things like the ChainstateManager.

As a part of this, we also get rid of some "static globals" and
move them as proper members (also with proper, well-defined life cycle)
into the CoinMintingModule; for instance, the cache of hashed blocks
and their timestamps, and the ChainstateManager reference used.