De-globalise ActiveChainManager.
ActiveChainManager is a cheap and simple class, which mainly holds
some references (thus is very cheap to construct/destruct) and whose
main use is structuring the logic for detaching blocks.

This change removes the GetActiveChainManager() method (which holds a
"static global" instance of the class), replacing it with direct
instantiation of the ActiveChainManager in the places where it is
needed.  This helps clarify the code and its interdependencies, and
also helps better ensure that the life cycles of instances are
harmonised to each other.