Pass ChainstateManager to ConnectBlock.
This refactors the calls to ConnectBlock and up in the hierarchy
(e.g. ActivateBestChain, AcceptBlockHeader) to explicitly pass
the ChainstateManager reference and CSporkManager down from a
higher level.

The main places that obtain the references from "globals" are now the
RPC functions as well as ProcessMessage in main.cpp (and some others
in init code and other places).