Introduce RPCContext instance.
This introduces the RPCContext class, and an associated global instance
(accessible through RPCContext::Get).  This class is empty for now, but
will in the future hold contextual information such as references to the
global chainstate, masternode module and others, for use by RPC method
handlers while the RPC server is running.

It will be difficult to explicitly pass references like these to the
RPC handlers, as that would require updating the entire RPC server
infrastructure and all RPC handler methods at once.  With this class, we
still lay the foundation for removing direct "global" accesses from
the RPC methods, and instead bundling them in the RPCContext class.  This
makes the use of global state from the RPC server clearer, and can be
used later on to fully remove any global access by simply providing the
RPCContext instance explicitly to handlers somehow.